# perf: speed up parsing by up to 60%