--- a/packages/prettier-plugin-apex/src/parser.ts
+++ b/packages/prettier-plugin-apex/src/parser.ts
@@ -382,8 +382,7 @@ function handleNodeLocation(
   }
   if (!node.loc) {
     delete node.loc;
-  }
-  if (node.loc && currentLocation) {
+  } else if (node.loc && currentLocation) {
     if (node.loc.startIndex > currentLocation.startIndex) {
       node.loc.startIndex = currentLocation.startIndex;
     } else {
@@ -448,10 +447,9 @@ function generateExtraMetadata(
   if (apexClass === APEX_TYPES.SEARCH || apexClass === APEX_TYPES.QUERY) {
     node.forcedHardline = node.loc.startLine !== node.loc.endLine;
   }
-
   // jorje parses all `if` and `else if` blocks into `ifBlocks`, so we add
   // `ifBlockIndex` into the node for handling code to differentiate them.
-  if (apexClass === APEX_TYPES.IF_ELSE_BLOCK) {
+  else if (apexClass === APEX_TYPES.IF_ELSE_BLOCK) {
     node.ifBlocks.forEach((ifBlock: jorje.IfBlock, index: number) => {
       (ifBlock as EnrichedIfBlock).ifBlockIndex = index;
     });