<!-- Please provide a brief summary of your changes: -->
Up to 60% faster parsing. The biggest improvement is for larger files.
The changes are around resolving the line indexes from character indexes and so most changes are in `resolvelineindexes`.
I implemented a binary search to find the line numbers for the given `startIndex` and `endIndex`

## Test numbers

**95 line Apex class**
before: 1087.75 ms
after: 1068.5 ms

**528 line Apex class**
before: 190 ms
after: 156.5 ms

**11_230 line Apex class**
before: 5710.25 ms
after: 3434.75 ms

## Images
Produced using `node --inspect-brk` and chrome profiling.
The Apex Class used was 11_230 lines long

**Before**
![prettier-apex-perf-before](https://github.com/user-attachments/assets/ab824444-afcf-43d5-8ec9-6ccbc033a4aa)

**After**
![prettier-apex-perf-after](https://github.com/user-attachments/assets/c0a34434-6466-465d-99f2-8b27faf7bebb)


<!-- Please ensure you’ve done all of these things (if applicable). -->
<!-- You can replace the `[ ]` with `[x]` to mark each task as done. -->

- [X] I’ve added tests to confirm my change works.

This was a perf improvement but I did testing and have attached before and after images of the flame chart as well as before and after numbers of time taken.

- [X] (If changing formatting output/API or CLI) I’ve added my changes to the `CHANGELOG.md` file in the `Unreleased` section.
- [X] I’ve read the [contributing guidelines](https://github.com/dangmai/prettier-plugin-apex/blob/master/CONTRIBUTING.md).