@dangmai 2024-10-06
Somehow I missed the notification for this PR! Thank you very much for taking a look into this 💯 , this part of the code hasn't been touched in a while and I know there's plenty of opportunities to make it faster.
There are currently some linting issues, do you mind fixing them?