@lukecotter 2024-10-07
Linting issues all fixed 👍🏻 

I had to switch from for..of back to forEach, due to the rule _"iterators/generators require regenerator-runtime, which is too heavyweight for this guide to allow them. Separately, loops should be avoided in favor of array iterations."_

In my testing for..of was faster but it was not a huge difference so happy to keep forEach if you would prefer.