@lukecotter 2024-10-08
Thanks for the merge. Looking forward to faster formatting 😀