@dangmai 2024-10-11
> > > Thanks for the merge. Looking forward to faster formatting 😀
> > 
> > 
> > There's some exciting development in store for you then 😉  I figured out a theoretical way to make the parsing a few orders of magnitude faster. Not ready to share it publicly yet, but it's coming.
> 
> Exciting, I'll keep an eye out! Is it still worth me experimenting with a new parser?

I'd hold off on that for a bit, unless you just want to poke around and see how Prettier Apex works 😄